move network config to proto
